+++ api-gateway/src/main/resources/application.yml
@@ -53,11 +53,11 @@ jwt:
 server:
   port: 8080
 
-management:
-  endpoints:
-    web:
-      exposure:
-        include: health, info, prometheus
+#management:
+#  endpoints:
+#    web:
+#      exposure:
+#        include: health, info, prometheus
 
 logging:
   level:
