+++ api-gateway/src/main/resources/logback-apigateway.xml
@@ -9,9 +9,15 @@
             <timestampPattern>yyyy-MM-dd' 'HH:mm:ss.SSS</timestampPattern>
         </layout>
     </appender>
+    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
 
     <root level="debug">
-        <appender-ref ref="LOGSTASH" />
+        <appender-ref ref="LOGSTASH"/>
+        <appender-ref ref="CONSOLE"/>
     </root>
 
 </configuration>
